Interface/backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
paho-mqtt>=2.0.0
